requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
transformers
torch
